chore: move casclient
