+++ app/cli/internal/action/artifact_download.go
@@ -25,7 +25,7 @@ import (
 	"path"
 	"time"
 
-	casclient "github.com/chainloop-dev/chainloop/app/cli/internal/casclient/grpc"
+	"github.com/chainloop-dev/chainloop/internal/casclient"
 	"github.com/jedib0t/go-pretty/v6/progress"
 	"google.golang.org/grpc"
