+++ app/cli/internal/action/artifact_upload.go
@@ -19,7 +19,7 @@ import (
 	"context"
 	"time"
 
-	casclient "github.com/chainloop-dev/chainloop/app/cli/internal/casclient/grpc"
+	"github.com/chainloop-dev/chainloop/internal/casclient"
 	"github.com/jedib0t/go-pretty/v6/progress"
 	"google.golang.org/grpc"
 )
