+++ internal/casclient/uploader.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package grpc
+package casclient
 
 import (
 	"bytes"
